+++ src/control_plane/engine/command_status.rs
@@ -1,0 +1,7 @@
+/// Current state of an Engine exec process.
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+#[non_exhaustive]
+pub(crate) enum CommandStatus {
+    Running,
+    Exited(i64),
+}